Cancel drop event to prevent browser navigation
Firefox attempts navigation to the domain represented by the task list
item text.

Once we've determined the drop event is a task list item dropping onto a
task list, cancel any default browser behavior and handle the event
locally. If the drop event is for something other than a list item,
ignore it and allow the browser or another upstream event listener to
handle it.